+++ comics_crafter/src/main/java/com/ahlquist/comics_crafter/repository/PrintingRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ahlquist.comics_crafter.model.Printing;
 //This is the Printing repository.
 @Repository
-public interface PrintingRepository extends JpaRepository<Printing, Integer>{
+public interface PrintingRepository extends JpaRepository<Printing, Long>{
 
 }
