+++ comics_crafter/src/main/java/com/ahlquist/comics_crafter/service/PaperService.java
@@ -10,6 +10,6 @@ import com.ahlquist.comics_crafter.model.Paper;
 public interface PaperService {
 	List<Paper> getAllPapers();
 	void savePaper(Paper paper);
-	Paper getPaperById(Integer paper_id);
-	void deletePaperById(Integer paper_id);
+	Paper getPaperById(Long paper_id);
+	void deletePaperById(Long paper_id);
 }
