+++ comics_crafter/src/main/java/com/ahlquist/comics_crafter/service/PrintingService.java
@@ -7,6 +7,6 @@ import com.ahlquist.comics_crafter.model.Printing;
 public interface PrintingService {
 	List<Printing> getAllPrinting();
 	void savePrinting(Printing printing);
-	Printing getPrintingById(Integer printing_id);
-	void deletePrintingById(Integer printing_id);
+	Printing getPrintingById(Long printing_id);
+	void deletePrintingById(Long printing_id);
 }
